+++ docs/changelog.rst
@@ -18,6 +18,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - A new option :opt:`strip_trailing_spaces` to optionally remove trailing
   spaces from lines when copying to clipboard.
 
+- A new option :opt:`tab_bar_min_tabs` to control how many tabs must be
+  present before the tab-bar is shown (:iss:`1382`)
+
 - Automatically check for new releases and notify when an update is available,
   via the system notification facilities. Can be controlled by
   :opt:`update_check_interval` (:iss:`1342`)